+++ src/utils/consts/index.ts
@@ -1,6 +1,6 @@
 export { navigationOptions } from './navigationOptions';
 export { IconName, Icons } from './icons';
-
+export { emailPattern, passwordValidationPatterns } from './regexPatterns';
 import { Product } from '@/types';
 import tailwind from '../../../tailwind.config';
 export const colors: string[] = Object.values(tailwind.theme.colors);
@@ -19,15 +19,6 @@ export const monthList = [
   'November',
   'December'
 ];
-
-export const passwordValidationPatterns = {
-  hasNumber: /\d/,
-  hasLowerCase: /[a-z]/,
-  isLong: /^.{8,}$/
-};
-
-// eslint-disable-next-line max-len
-export const emailPattern = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
 
 export const authErrors = {
   'auth/invalid-email': {
